# studyindexer/app/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "StudyIndexer API",
    "description": "Document Processing and Semantic Search Service\n\nFeatures:\n* Document upload and processing (PDF, DOCX, TEXT, MARKDOWN)\n* Semantic search across documents\n* Course-specific document collections\n* Role-based access control\n* Document metadata management",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CollectionType": {
        "type": "string",
        "enum": ["general", "course", "personal", "faq"]
      },
      "BaseResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "default": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          }
        }
      },
      "ErrorResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponse"
          },
          {
            "type": "object",
            "properties": {
              "success": {
                "type": "boolean",
                "default": false
              },
              "error": {
                "$ref": "#/components/schemas/ErrorDetail"
              }
            }
          }
        ]
      },
...
      "SearchQuery": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "minLength": 3,
            "maxLength": 1000
          },
          "page": {
            "type": "integer",
            "minimum": 1,
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "minimum": 1,
            "maximum": 50,
            "default": 10
          },
          "filters": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "collection": {
            "$ref": "#/components/schemas/CollectionType",
            "default": "general"
          },
          "min_score": {
            "type": "number",
            "minimum": 0.0,
            "maximum": 1.0,
            "default": 0.0
          }
        },
        "required": ["text"]
      },
      "SearchResult": {
        "type": "object",
        "properties": {
          "document_id": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "minimum": 0.0,
            "maximum": 1.0
          },
          "content": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "page_number": {
            "type": "integer",
            "minimum": 1,
            "nullable": true
          },
          "position": {
            "type": "object",
            "nullable": true,
            "properties": {
              "chunk": {
                "type": "integer",
                "nullable": true
              },
              "start": {
                "type": "integer",
                "nullable": true
              },
              "end": {
                "type": "integer",
                "nullable": true
              }
            }
          },
          "highlight": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "PaginationMetadata": {
        "type": "object",
        "properties": {
          "current_page": {
            "type": "integer",
            "minimum": 1
          },
          "page_size": {
            "type": "integer",
            "minimum": 1
          },
          "total_pages": {
            "type": "integer",
            "minimum": 0
          },
          "total_results": {
            "type": "integer",
            "minimum": 0
          },
          "has_next": {
            "type": "boolean"
          },
          "has_previous": {
            "type": "boolean"
          }
        }
      },
      "SearchResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponse"
          },
          {
            "type": "object",
            "properties": {
              "results": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SearchResult"
                }
              },
              "pagination": {
                "$ref": "#/components/schemas/PaginationMetadata"
              },
              "query_time_ms": {
                "type": "number",
                "minimum": 0
              },
              "collection": {
                "type": "string"
              },
              "filters_applied": {
                "type": "object",
                "nullable": true,
                "additionalProperties": true
              }
            }
          }
        ]
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/search": {
      "post": {
        "summary": "Search documents",
        "description": "Search across documents using semantic search",
        "tags": ["search"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/similar/{document_id}": {
      "get": {
        "summary": "Find similar documents",
        "description": "Find documents similar to a given document",
        "tags": ["search"],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "description": "Document ID to find similar documents for",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of similar documents to return",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "name": "min_score",
            "in": "query",
            "description": "Minimum similarity score (0-1)",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 1,
              "default": 0.5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Similar documents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Document not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
} 